+++ src/core/runtime/detail/partition_manager.cc
@@ -239,7 +239,7 @@ bool PartitionManager::use_complete_tiling(const tuple<std::uint64_t>& extents,
   constexpr auto MAX_TILES_HEURISTIC  = 256;
   constexpr auto MAX_PIECES_HEURISTIC = 16;
   const auto num_tiles                = (extents / tile_shape).volume();
-  const auto num_pieces               = Runtime::get_runtime()->get_machine().count();
+  const auto num_pieces = static_cast<std::uint64_t>(Runtime::get_runtime()->get_machine().count());
   return num_tiles <= MAX_TILES_HEURISTIC || num_tiles <= MAX_PIECES_HEURISTIC * num_pieces;
 }
