+++ tests/cpp/integration/test_task_store.cc
@@ -232,7 +232,7 @@ void auto_task_normal_input(const legate::LogicalArray& array, std::uint32_t ind
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + array.dim());
+  auto task = runtime->create_task(context, static_cast<std::int64_t>(SIMPLE_TASK) + array.dim());
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::INPUT);
   constexpr auto store_type = static_cast<std::int32_t>(StoreType::NORMAL_STORE);
@@ -255,7 +255,7 @@ void auto_task_normal_output(const legate::LogicalArray& array, std::uint32_t in
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + array.dim());
+  auto task = runtime->create_task(context, static_cast<std::int64_t>(SIMPLE_TASK) + array.dim());
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::OUTPUT);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::NORMAL_STORE);
@@ -283,7 +283,7 @@ void auto_task_normal_reduction(const legate::LogicalArray& array, std::uint32_t
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + array.dim());
+  auto task = runtime->create_task(context, static_cast<std::int64_t>(SIMPLE_TASK) + array.dim());
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::REDUCTION);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::NORMAL_STORE);
@@ -319,7 +319,7 @@ void auto_task_unbound_input(const legate::LogicalArray& array, std::uint32_t in
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + array.dim());
+  auto task = runtime->create_task(context, static_cast<std::int64_t>(SIMPLE_TASK) + array.dim());
 
   switch (index) {
     case 0: EXPECT_THROW(task.add_input(array), std::invalid_argument); break;
@@ -335,7 +335,7 @@ void auto_task_unbound_output(const legate::LogicalArray& array, std::uint32_t i
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + array.dim());
+  auto task = runtime->create_task(context, static_cast<std::int64_t>(SIMPLE_TASK) + array.dim());
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::OUTPUT);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::UNBOUND_STORE);
@@ -359,7 +359,7 @@ void auto_task_unbound_reduction(const legate::LogicalArray& array, std::uint32_
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + array.dim());
+  auto task = runtime->create_task(context, static_cast<std::int64_t>(SIMPLE_TASK) + array.dim());
 
   constexpr auto red_op = legate::ReductionOpKind::ADD;
 
@@ -390,7 +390,8 @@ void manual_task_normal_input(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::INPUT);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::NORMAL_STORE);
@@ -419,7 +420,8 @@ void manual_task_normal_output(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::OUTPUT);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::NORMAL_STORE);
@@ -451,7 +453,8 @@ void manual_task_normal_reduction(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::REDUCTION);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::NORMAL_STORE);
@@ -494,7 +497,8 @@ void manual_task_unbound_input(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   switch (index) {
     case 0: EXPECT_THROW(task.add_input(store), std::invalid_argument); break;
@@ -508,7 +512,8 @@ void manual_task_unbound_output(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::OUTPUT);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::UNBOUND_STORE);
@@ -534,7 +539,8 @@ void manual_task_unbound_reduction(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto red_op = legate::ReductionOpKind::ADD;
 
@@ -555,7 +561,8 @@ void manual_task_scalar_input(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::INPUT);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::SCALAR_STORE);
@@ -583,7 +590,8 @@ void manual_task_scalar_output(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::OUTPUT);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::SCALAR_STORE);
@@ -615,7 +623,8 @@ void manual_task_scalar_reduction(const legate::LogicalStore& store,
 {
   auto runtime = legate::Runtime::get_runtime();
   auto context = runtime->find_library(library_name);
-  auto task    = runtime->create_task(context, SIMPLE_TASK + store.dim(), launch_shape);
+  auto task    = runtime->create_task(
+    context, static_cast<std::int64_t>(SIMPLE_TASK) + store.dim(), launch_shape);
 
   constexpr auto mode       = static_cast<std::uint32_t>(TaskDataMode::REDUCTION);
   constexpr auto store_type = static_cast<std::uint32_t>(StoreType::SCALAR_STORE);
@@ -1002,7 +1011,7 @@ TEST_F(Runtime, CreateTaskInvalid)
                std::out_of_range);
 
   EXPECT_THROW(static_cast<void>(runtime->create_task(context,
-                                                      SIMPLE_TASK + 1,
+                                                      static_cast<std::int64_t>(SIMPLE_TASK) + 1,
                                                       legate::tuple<std::uint64_t>{
                                                         0,
                                                       })),
